requirements.txt
requests>=2.31,<3
mysql-connector-python>=8.0
